# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Blue Sky Analytics Assignment 2",
      "version": "1.0.0"
    },
    "paths": {
...
      "/country/{id}": {
        "get": {
          "parameters": [
            {
                "name": "id",
                "in": "path",
                "description": "Country Id",
                "required": true,
                "type": "string",
                "example": "6137d8aaecfca86db07d6035"
            },
            {
              "name": "startDate",
              "in": "query",
              "description": "Start Year for the data",
              "required": true,
              "type": "string",
              "example": "1999"

            },
            {
              "name": "category",
              "in": "query",
              "description": "Category of data",
              "required": true,
              "type": "string",
              "example": "n2o,co2"


            },
            {
              "name": "endDate",
              "in": "query",
              "description": "End Year for the data",
              "required": false,
              "type": "string",
              "example": "2014"

            }
        ],
          "operationId": "getSpecificCountryDetail",
          "summary": "Show list of values according to parameters for specific country",
          "responses": {
            "200": {
              "description": "200 response",
              "content": {
                "application/json": {
                  "examples": {
                    "Country Data": {
                      "value": [
                        {
                            "_id": "6137e7e1de9474a5abf64e2a",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2010,
                            "value": 97035.0827603655,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e2d",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2009,
                            "value": 103969.78429385,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e30",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2008,
                            "value": 110255.765521907,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e33",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2007,
                            "value": 114127.554270902,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e36",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2006,
                            "value": 111777.743714104,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e39",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2005,
                            "value": 113219.822395958,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e3c",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2004,
                            "value": 109571.060909579,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e3f",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2003,
                            "value": 109126.567554916,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e42",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2002,
                            "value": 105051.466199101,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e45",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2001,
                            "value": 105407.307598309,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e1de9474a5abf64e48",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2000,
                            "value": 103019.664690088,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e4b",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1999,
                            "value": 97979.7783523224,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e4e",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1998,
                            "value": 98663.1941625214,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e51",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1997,
                            "value": 93839.8699189575,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e54",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1996,
                            "value": 89134.4253965711,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e57",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1995,
                            "value": 86980.6457725937,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e5a",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1994,
                            "value": 86425.8565118787,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e5d",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1993,
                            "value": 84263.0078976228,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e60",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1992,
                            "value": 84949.0258298748,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e63",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1991,
                            "value": 83390.3546805713,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137e7e2de9474a5abf64e66",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1990,
                            "value": 83403.7199982216,
                            "category": "co2",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a518f1",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2010,
                            "value": 5526.82836121731,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a518f4",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2009,
                            "value": 5325.95590579888,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a518f7",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2008,
                            "value": 5610.97455334186,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a518fa",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2007,
                            "value": 5825.36161258537,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a518fd",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2006,
                            "value": 5730.83250457993,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a51900",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2005,
                            "value": 5882.94951366643,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a51903",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2004,
                            "value": 6054.52883831513,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb001d2bc408a3a51906",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2003,
                            "value": 6050.25253250507,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51909",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2002,
                            "value": 6172.35756103089,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a5190c",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2001,
                            "value": 6207.72807311416,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a5190f",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 2000,
                            "value": 6346.04941154567,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51912",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1999,
                            "value": 6604.32412668112,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51915",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1998,
                            "value": 6621.5570833371,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51918",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1997,
                            "value": 6686.38644285652,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a5191b",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1996,
                            "value": 6846.91196407554,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a5191e",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1995,
                            "value": 6669.80922700312,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51921",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1994,
                            "value": 6480.19419955965,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51924",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1993,
                            "value": 6564.37569257449,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb011d2bc408a3a51927",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1992,
                            "value": 7129.88046160346,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb021d2bc408a3a5192a",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1991,
                            "value": 7269.55164996042,
                            "category": "n2o",
                            "__v": 0
                        },
                        {
                            "_id": "6137eb021d2bc408a3a5192d",
                            "countryID": "6137d8aaecfca86db07d6043",
                            "year": 1990,
                            "value": 7402.64281402762,
                            "category": "n2o",
                            "__v": 0
                        }
                    ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }